% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{all_transition_network_graph}
\alias{all_transition_network_graph}
\title{List of the structure of all the transition networks}
\usage{
all_transition_network_graph(tvdbn.fit)
}
\arguments{
\item{tvdbn.fit}{A fitted Time-varying DBN of type \code{tvdbn.fit}.}
}
\value{
A list \code{tvdbn} containing the structure (type \code{tvdbn}) of all the
transition networks
}
\description{
Get all the transition network structures of the Time-varying DBN
}
