% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tvdbn_class.R
\name{get_time_points}
\alias{get_time_points}
\title{Get the number of time points of a Time-Varying DBN}
\usage{
get_time_points(tvdbn)
}
\arguments{
\item{tvdbn}{The Time-varying DBN. Can be an object of type \code{tvdbn} or \code{tvdbn.fit}}
}
\value{
The number of time instants of the Time-Varying DBN.
}
\description{
Get the number of time points of time instants of the Time-Varying DBN.
}
