% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tvdbn_class.R
\name{get_variables}
\alias{get_variables}
\title{Get the template variables}
\usage{
get_variables(tvdbn)
}
\arguments{
\item{tvdbn}{The Time-varying DBN. Can be an object of type \code{tvdbn} or \code{tvdbn.fit}}
}
\value{
The number of template variables of the Time-Varying DBN.
}
\description{
Get the template variables of our TV-DBN, i.e. the variables not
instantiated in each time point.
}
