% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{graph}
\alias{graph}
\title{Structure of a Time-varying DBN}
\usage{
graph(tvdbn.fit)
}
\arguments{
\item{tvdbn.fit}{A fitted Time-varying DBN of type \code{tvdbn.fit}}
}
\value{
The structure of the TV-DBN, type = \code{tvdbn}
}
\description{
Get the structure of a Time-varying DBN given a fitted Time-varying DBN
}
