% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xai.R
\name{hamming_changes}
\alias{hamming_changes}
\title{The hamming distances accross the transition networks.}
\usage{
hamming_changes(tvdbn.fit)
}
\arguments{
\item{tvdbn.fit}{A fitted Time-varying DBN of type \code{tvdbn.fit}.}
}
\value{
A vector with the Hamming distance between adyadcents transition networks.
Each position i represents the Hamming distance between the \code{i-th} and \code{i+1-th}
transition network.
}
\description{
Get the difference between all the adyacent transition networks
of a given Time-varying Dynamic Bayesian Network
}
