% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_time_name}
\alias{remove_time_name}
\title{Get the name of a variable and the time point from a time name}
\usage{
remove_time_name(time_name)
}
\arguments{
\item{time_name}{The name of a variable with a time point correctly appended.
Be careful to supply a string with the correct format}
}
\value{
A vector with two elements:
\enumerate{
\item The name of the variable (without the time appended)
\item The time point (be careful, since it the type is \code{character})
}
}
\description{
Get the name of a variable and the time point from a time name.
}
