% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{subgraph}
\alias{subgraph}
\title{Structure of a node subset of a Time-varying DBN}
\usage{
subgraph(tvdbn.fit, nodes)
}
\arguments{
\item{tvdbn.fit}{A fitted Time-varying DBN of type \code{tvdbn.fit}}

\item{nodes}{The nodes whose structure will be obtained.}
}
\value{
The structure of the TV-DBN, type = \code{tvdbn}
}
\description{
Get the structure of a Time-varying DBN given a fitted Time-varying DBN,
considering only a subset of nodes of the network
}
