% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{time_name}
\alias{time_name}
\title{Construct the time name of a variable}
\usage{
time_name(name, t)
}
\arguments{
\item{name}{Name of the variable}

\item{t}{Time instant (integer)}
}
\value{
The name of the variable with the time point correctly appended.
}
\description{
Construct the time name of a variable (i.e. the name of the variable with the time point correctly appended)
}
