% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{transition_network}
\alias{transition_network}
\title{Transition network in a time point}
\usage{
transition_network(tvdbn.fit, time)
}
\arguments{
\item{tvdbn.fit}{A fitted Time-varying DBN of type \code{tvdbn.fit}}
}
\value{
The structure of the TV-DBN, type = \code{tvdbn}
}
\description{
Get the transition network of a certain time instant.
}
