% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{transition_network_graph}
\alias{transition_network_graph}
\title{Structure of the transition network in a time point}
\usage{
transition_network_graph(tvdbn.fit, time)
}
\arguments{
\item{tvdbn.fit}{A fitted Time-varying DBN of type \code{tvdbn.fit}.}

\item{time}{The transition network to obtain.}
}
\value{
The structure of the transition network of the TV-DBN in
the given time instant.
}
\description{
Get the transition network of a certain time instant.
}
