% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{transition_network_markovian_order}
\alias{transition_network_markovian_order}
\title{Markovian order of a transition network}
\usage{
transition_network_markovian_order(trans_network)
}
\arguments{
\item{trans_network}{A Time-varying DBN structure, type \code{tvdbn}.}
}
\value{
The Markovian order of the transition network.
}
\description{
Get the Markovian order of a transition network. In
other words, the number of past instant from which the transition
network model is dependent.
}
