% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{transition_network_normalize_name}
\alias{transition_network_normalize_name}
\title{Normalize the name of the time variables of a TN}
\usage{
transition_network_normalize_name(trans_network)
}
\arguments{
\item{tvdbn.fit}{A fitted Time-varying DBN of type \code{tvdbn}.}
}
\value{
The transition network with the naming normalized
}
\description{
Normalize the name of the time variables of a Transition Network.
This allows comparing them using the Hamming distance
}
